Octoml Config files
With this commit we have this structure:

/models
  /tensorflow_models
    /style_model.tar.gz
    /octoml.yaml
    /accelerated
      /octoml.yaml
  /onnx_models
    /gpt_model.onnx
    /octoml.yaml
    /accelerated
      /octoml.yaml

How to package
For local pacakging for the model you can cd to e.g.
models/onnx_models and run `octoml package`.

For express packaging do the same as local but use `octoml package -e`

To do full acceleration with hardware cd to
models/onnx_models/acceleration and run `octoml package -a.`

We may end up with another layer e.g.
models/onnx_runtime/model_name/octoml.yaml but we're not there yet